refactor(transfer): decompose receiver/file_list.rs (SPL-27 #2455)
Split the 1195-line monolith into a `file_list/` module with five
focused submodules (each under 650 lines):

- `mod.rs` - module surface and re-exports
- `receive.rs` - `receive_file_list`, `receive_extra_file_lists`,
  `publish_segment_to_delete_pipeline`, `incremental_file_list_receiver`
- `id_lists.rs` - UID/GID name mapping reception (unix + non-unix)
- `sanitize.rs` - trust-gated path validation against traversal
- `hardlinks.rs` - post-sort `match_hard_links` and pre-30 normalization
- `incremental.rs` - the streaming `IncrementalFileListReceiver`

Zero public API change: `IncrementalFileListReceiver` continues to be
re-exported via `receiver::mod` and all `pub(in crate::receiver)` or
`pub(crate)` items keep their original visibility scopes.
